+++ edsa_recommender.py
@@ -1,29 +1,20 @@
 """
-
     Streamlit webserver-based Recommender Engine.
-
     Author: Explore Data Science Academy.
-
     Note:
     ---------------------------------------------------------------------
     Please follow the instructions provided within the README.md file
     located within the root of this repository for guidance on how to use
     this script correctly.
-
     NB: !! Do not remove/modify the code delimited by dashes !!
-
     This application is intended to be partly marked in an automated manner.
     Altering delimited code may result in a mark of 0.
     ---------------------------------------------------------------------
-
     Description: This file is used to launch a minimal streamlit web
 	application. You are expected to extend certain aspects of this script
     and its dependencies as part of your predict project.
-
 	For further help with the Streamlit framework, see:
-
 	https://docs.streamlit.io/en/latest/
-
 """
 # Streamlit dependencies
 import streamlit as st
@@ -31,29 +22,26 @@ import streamlit as st
 # Data handling dependencies
 import pandas as pd
 import numpy as np
+import re
+from PIL import Image
+#Plots
+import seaborn as sns
+import matplotlib.style as style 
+sns.set(font_scale=1)
+import matplotlib.pyplot as plt
+from datetime import date
 
 # Custom Libraries
 from utils.data_loader import load_movie_titles
 from recommenders.collaborative_based import collab_model
 from recommenders.content_based import content_model
 st.markdown("""<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">""",unsafe_allow_html=True)
-import matplotlib.pyplot as plt
-from datetime import date
-import seaborn as sns
-sns.set(font_scale=1)
 
-# Custom Libraries
-from utils.data_loader import load_movie_titles
-from recommenders.collaborative_based import collab_model
-from recommenders.content_based import content_model
-
-# Setting Data Path
 data_path = '../unsupervised_data/unsupervised_movie_data/'
-
 # Data Loading
 title_list = load_movie_titles('resources/data/movies.csv')
 
-# EDA Content
+#Movies_list EDA data
 titles = ['The Shawshank Redemption (1994)', 'Pulp Fiction (1994)', 'Forrest Gump (1994)',
           'The Silence of The Lambs (1991)',
           'The MATRIX (1999)', 'Star Wars: Episode IV - A New Hope (1977)', 'Schindler\'s List (1993)',
@@ -89,12 +77,13 @@ descriptions = ['Genre: Drama  \nRealese year:1994  \nRuntime: 2h 22min  \nStory
                 'Genre: Action, Adventure, Drama  \nRuntime: 2h.58min \nStoryline:  \nAn ancient Ring thought lost for centuries has been found, and through a strange twist of fate has been given to a small Hobbit named Frodo. When Gandalf discovers the Ring is in fact the One Ring of the Dark Lord Sauron, Frodo must make an epic quest to the Cracks of Doom in order to destroy it. However, he does not go alone. He is joined by Gandalf, Legolas the elf, Gimli the Dwarf, Aragorn, Boromir, and his three Hobbit friends Merry, Pippin, and Samwise. Through mountains, snow, darkness, forests, rivers and plains, facing evil and danger at every corner the Fellowship of the Ring must go. Their quest to destroy the One Ring is the only hope for the end of the Dark Lords reign.',
                 'Genre: Action, Adventure  \nRuntime: 1h.55m  \nRatings: 4.1 \nStoryline:  \nThe year is 1936. An archeology professor named Indiana Jones is venturing in the jungles of South America searching for a golden statue. Unfortunately, he sets off a deadly trap but miraculously escapes. Then, Jones hears from a museum curator named Marcus Brody about a biblical artifact called The Ark of the Covenant, which can hold the key to humanely existence. Jones has to venture to vast places such as Nepal and Egypt to find this artifact. However, he will have to fight his enemy Rene Belloq and a band of Nazis in order to reach it.']
 
+
 # App declaration
 def main():
 
     # DO NOT REMOVE the 'Recommender System' option below, however,
     # you are welcome to add more options to enrich your app.
-    page_options = ["Recommender System","Exploratory Data Analysis","NEXTFLIX Special","Solution Overview","About us"]
+    page_options = ['Recommender System','Exploratory Data Analysis','What would you like to find?','Solution Overview','About us']
 
     # -------------------------------------------------------------------
     # ----------- !! THIS CODE MUST NOT BE ALTERED !! -------------------
@@ -105,6 +94,9 @@ def main():
         st.write('# Movie Recommender Engine')
         st.write('### EXPLORE Data Science Academy Unsupervised Predict')
         st.image('resources/imgs/Image_header.png',use_column_width=True)
+        ratings_df = pd.read_csv('resources/data/ratings.csv')
+        movies_df = pd.read_csv('resources/data/movies.csv')
+        ratings_new = ratings_df.merge(movies_df, on='movieId', how='left')
         # Recommender System algorithm selection
         sys = st.radio("Select an algorithm",
                        ('Content Based Filtering',
@@ -149,7 +141,7 @@ def main():
     # -------------------------------------------------------------------
 
     # ------------- SAFE FOR ALTERING/EXTENSION -------------------
-    if page_selection == "Exploratory Data Analysis":
+    if page_selection == 'Exploratory Data Analysis':
         st.markdown("""
                     <div class="w3-display-container" style="margin-bottom:15px">
                     <img src="https://www.deluxetheatre.co.nz/images/index/slider/slider_1_bg.jpg" style="width:100%">
@@ -158,11 +150,8 @@ def main():
                     </div>
                     </div>
                     """, unsafe_allow_html=True)
-        sub_pages = ['WordClouds', 'Rating Distributions', 'Greatest Hits', 'Production costs over time']
-        st.sidebar.markdown("""
-                            * Hey, pick one from the options below to view the results of our EDA 
-                            """)
-        sub_page = st.sidebar.radio('EDA options', sub_pages)
+        sub_pages = ['WordClouds','Rating Distributions','Greatest Hits']
+        sub_page = st.selectbox('Choose EDA options:',sub_pages)
         if sub_page == 'WordClouds':
             st.markdown("""
                         <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue">
@@ -171,15 +160,13 @@ def main():
                         </div>
                         """, unsafe_allow_html=True)
             if st.checkbox('view actors wordcloud'):
-                st.image('resources/imgs/EDA_imgs/actors_wordcloud.png', caption='Actors WordCloud')
+                st.image('resources/imgs/EDA_imgs/actors_wordcloud.png',caption='Actors WordCloud')
                 st.subheader('Observations:')
-                st.markdown(
-                    '''* Our intuition was right, immediately we see some big names in the film industry jumping out at us.  \n* Stephen King and Shakespeare? Well.. Its more likely that these films were based on their work.  \n* Woody Allen is an American director, writer, actor, and comedian whose career spans more than six decades and multiple Academy Award-winning movies.  \n* Tyler Perry has directed some amazing films like Acrimony and the Madea films. He now has his own motion picture studio, 'Tyler Perry Studios'.  \n* Luc Paul Maurice Besson is a French film director, screenwriter, and producer. He directed or produced the films Subway, The Big Blue, and La Femme Nikita. Besson is associated with the Cinéma du look film movement.''')
+                st.markdown('''* Our intuition was right, immediately we see some big names in the film industry jumping out at us.  \n* Stephen King and Shakespeare? Well.. Its more likely that these films were based on their work.  \n* Woody Allen is an American director, writer, actor, and comedian whose career spans more than six decades and multiple Academy Award-winning movies.  \n* Tyler Perry has directed some amazing films like Acrimony and the Madea films. He now has his own motion picture studio, 'Tyler Perry Studios'.  \n* Luc Paul Maurice Besson is a French film director, screenwriter, and producer. He directed or produced the films Subway, The Big Blue, and La Femme Nikita. Besson is associated with the Cinéma du look film movement.''')
             if st.checkbox('view directors wordcloud'):
-                st.image('resources/imgs/EDA_imgs/directors_wordcloud.png', caption='Directors WordCloud')
+                st.image('resources/imgs/EDA_imgs/directors_wordcloud.png',caption='Directors WordCloud')
                 st.subheader('Observations:')
-                st.markdown(
-                    '''* Tom Hanks: Aerican actor and filmmaker. Known for both his comedic and dramatic roles, Hanks is one of the most popular and recognizable film stars worldwide, and is widely regarded as an American cultural icon.  \n* Ben Stiller: Throughout his career he has written, starred in, directed, or produced more than 50 films including The Secret Life of Walter Mitty, Zoolander, The Cable Guy and There's Something About Mary.  \n* Eddie Murphy and Chris Rock are both successful comedians turned actors who have starred in numerous films throughout their careers.''')
+                st.markdown('''* Tom Hanks: Aerican actor and filmmaker. Known for both his comedic and dramatic roles, Hanks is one of the most popular and recognizable film stars worldwide, and is widely regarded as an American cultural icon.  \n* Ben Stiller: Throughout his career he has written, starred in, directed, or produced more than 50 films including The Secret Life of Walter Mitty, Zoolander, The Cable Guy and There's Something About Mary.  \n* Eddie Murphy and Chris Rock are both successful comedians turned actors who have starred in numerous films throughout their careers.''')
         if sub_page == 'Rating Distributions':
             st.markdown("""
                         <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue">
@@ -187,12 +174,10 @@ def main():
                         <p>Take a look at the <i>Rating Distributions</i></p>
                         </div>
                         """, unsafe_allow_html=True)
-            st.image(['resources/imgs/EDA_imgs/movielens_distribution_donut.png',
-                      'resources/imgs/EDA_imgs/average_distributions.png'], width=400,
-                     caption=['', 'User rating distributions'])
+            st.image(['resources/imgs/EDA_imgs/movielens_distribution_donut.png','resources/imgs/EDA_imgs/average_distributions.png'],width=400,
+                     caption=['','User rating distributions'])
             st.subheader('Observations:')
-            st.markdown(
-                '''* 4 Star ratings make up the largest portion of ratings in the MovieLens dataset, accounting for 26.5% of the overall ratings.  \n* 5 star ratings make up 14.5% of the overall ratings (3rd largest portion).  \n* 0.5 star ratings account for the smallest portion of the ratings at a mere 1.6%.  \n* Most of the movies have received less than 2500 ratings. While the number of movies having more than 5000 ratings is very low.  \n* The ratings follow a normal distribution that is slightly skewed to the right. It seems like users are generally generous with thier ratings.''')
+            st.markdown('''* 4 Star ratings make up the largest portion of ratings in the MovieLens dataset, accounting for 26.5% of the overall ratings.  \n* 5 star ratings make up 14.5% of the overall ratings (3rd largest portion).  \n* 0.5 star ratings account for the smallest portion of the ratings at a mere 1.6%.  \n* Most of the movies have received less than 2500 ratings. While the number of movies having more than 5000 ratings is very low.  \n* The ratings follow a normal distribution that is slightly skewed to the right. It seems like users are generally generous with thier ratings.''')
         if sub_page == 'Greatest Hits':
             st.markdown("""
                         <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue">
@@ -200,9 +185,8 @@ def main():
                         <p>Here are the <i>Greatest Hits</i>, use the 'next' and 'previous' buttons to navigate</p>
                         </div>
                         """, unsafe_allow_html=True)
-            # Next Button
-            button = 'button'  # Initialise program only
-
+            #Next Button
+            button = 'button' #Initialise program only
             def next_prev(button):
                 next_counter = pd.read_csv('resources/imgs/thumbnails/next_button.csv')['value']
                 next_value = next_counter[0]
@@ -220,17 +204,16 @@ def main():
                         next_value = next_value - 1
                     next_counter[0] = next_value
                     next_counter.to_csv('resources/imgs/thumbnails/next_button.csv')
-                return (next_value)
-
-            st.subheader('{}. {}'.format(next_prev(button) + 1, titles[next_prev(button)]))
-            st.image(images[next_prev(button)], width=500)
+                return(next_value)
+            st.subheader('{}. {}'.format(next_prev(button)+1,titles[next_prev(button)]))
+            st.image(images[next_prev(button)],width=500)
             st.markdown(descriptions[next_prev(button)])
             if st.button('next'):
                 next_prev('next')
             if st.button('previous'):
                 next_prev('previous')
-
-    if page_selection == "NEXTFLIX Special":
+                
+    if page_selection == 'What would you like to find?':
         st.markdown("""
                     <div class="w3-display-container" style="margin-bottom:15px">
                     <img src="https://png.pngtree.com/thumb_back/fw800/back_our/20190619/ourmid/pngtree-european-business-movie-poster-board-vector-background-material-image_134683.jpg" style="width:100%">
@@ -242,50 +225,49 @@ def main():
                     <p style="padding-top: 10px;">Let's have some fun... Find your NEXTFLIX</p>
                     </div>
                     """, unsafe_allow_html=True)
-        drop_down_listings = st.selectbox('What would you like to know:',
-                                          ['What\'s in a genre?', 'Hottest Movie releases', 'Movie Search'])
+        st.subheader('Nextflix') 
+        drop_down_listings = st.selectbox('What would you like to know:',['What\'s in a genre?','Find Hottest Movie releases','Movie Search'])
         if drop_down_listings == 'What\'s in a genre?':
-            genres_setlist = ['Action', 'Adventure', 'Animation',
-                              'Children', 'Comedy',
-                              'Crime', 'Documentary',
-                              'Drama', 'Fantasy', 'Horror', 'Mystery',
-                              'Romance', 'Sci-fi',
-                              'Thriller', 'War', 'Western']
-            genres = st.multiselect('Pick your favourie genre(s) for some fun facts', genres_setlist)
-            #        st.write(genres)
+            genres_setlist = ['Action','Adventure','Animation',
+                              'Children','Comedy',
+                              'Crime','Documentary',
+                              'Drama','Fantasy','Horror','Mystery',
+                              'Romance','Sci-fi',
+                              'Thriller','War','Western']
+            genres = st.multiselect('Pick your favourie genre(s) for some fun facts',genres_setlist)
+
             if len(genres) > 0:
-                def genre_count(filename, list1):
+                def genre_count(filename,list1):
                     '''Plots the distribution of genres in the movies dataset'''
-                    filename = data_path + str(filename)
-                    chunks = pd.read_csv(filename, chunksize=10000)
+                    filename = data_path+str(filename)
+                    chunks = pd.read_csv(filename,chunksize=10000)
                     data = pd.DataFrame()
                     count = 0
                     dict_genres = {}
                     for chunk in chunks:
-                        chunk_genres = ','.join([genres.replace('|', ',') for genres in chunk.genres]).split(',')
+                        chunk_genres = ','.join([genres.replace('|',',') for genres in chunk.genres]).split(',')
                         chunk_genres = [item for item in chunk_genres if item in list1]
                         for genre in chunk_genres:
                             if genre in dict_genres:
-                                dict_genres[genre] += 1
+                                dict_genres[genre]+=1
                             else:
-                                dict_genres[genre] = 1
-                    sorted_dict = sorted(dict_genres.items(), key=lambda x: x[1], reverse=True)
+                                dict_genres[genre]=1
+                    sorted_dict = sorted(dict_genres.items(), key=lambda x: x[1],reverse=True)
                     genre, frequency = zip(*sorted_dict)
-                    plt.figure(figsize=(10, 5))
-                    freq_plot = sns.barplot(x=frequency, y=list(genre), palette='pastel')
+                    plt.figure(figsize=(10,5))
+                    freq_plot = sns.barplot(x = frequency,y = list(genre),palette='pastel')
                     freq_plot.set(title='Number of movies in genre',
-                                  xlabel='Genre_count', ylabel='Genre')
+                                  xlabel='Genre_count',ylabel='Genre')
                     plt.show()
                     return (freq_plot)
-
-                genre_count_figure = genre_count('movies.csv', genres).figure
+                genre_count_figure = genre_count('movies.csv',genres).figure
                 if st.checkbox('show genre counts in dataset'):
                     st.write(genre_count_figure)
 
-        if drop_down_listings == 'Hottest Movie releases':
-            current_year = date.today().year
+        if drop_down_listings =='Find Hottest Movie releases':
+            current_year=date.today().year
             min_year = 1970
-            add_slider = st.slider('Choose year range:', min_year, current_year, (min_year, current_year), step=1)
+            add_slider = st.slider('Choose year range:',min_year, current_year,(min_year,current_year), step=1)
             if st.button('View highest rated in time frame'):
                 def average_by_year(filename):
                     a = add_slider[0]
@@ -293,122 +275,110 @@ def main():
                     counter = 0
                     data_f1 = pd.DataFrame()
                     data = pd.DataFrame()
-                    chunks_1 = pd.read_csv('../unsupervised_data/unsupervised_movie_data/movies.csv', chunksize=50000)
+                    chunks_1 = pd.read_csv('../unsupervised_data/unsupervised_movie_data/movies.csv',chunksize=50000)
                     for chunk in chunks_1:
-                        chunk['release_year'] = chunk.title.map(lambda x: re.findall('\d\d\d\d', x))
+                        chunk['release_year'] = chunk.title.map(lambda x: re.findall('\d\d\d\d',x))
                         chunk.release_year = chunk.release_year.apply(lambda x: np.nan if not x else int(x[-1]))
-                        chunk = chunk[(chunk['release_year'] >= a) & (chunk['release_year'] < b + 1)].drop(['genres'],
-                                                                                                           axis=1)
-                        data_f1 = pd.concat([chunk, data_f1])
+                        chunk = chunk[(chunk['release_year']>=a)&(chunk['release_year']<b+1)].drop(['genres'],axis=1)
+                        data_f1 = pd.concat([chunk,data_f1])
                     id_in_movies = data_f1.movieId
 
-                    chunks_2 = pd.read_csv('../unsupervised_data/unsupervised_movie_data/train.csv', chunksize=50000)
+                    chunks_2 = pd.read_csv('../unsupervised_data/unsupervised_movie_data/train.csv',chunksize=50000)
                     for chunk in chunks_2:
-                        identical = chunk[chunk['movieId'].isin(id_in_movies)].drop(['userId', 'timestamp'], axis=1)
-                        data = pd.concat([identical, data])
-
+                        identical = chunk[chunk['movieId'].isin(id_in_movies)].drop(['userId','timestamp'],axis=1)
+                        data = pd.concat([identical,data])
                     def top_movies_list(df=data):
                         m = df.movieId.value_counts().quantile(0.8)
                         vote_count = df.movieId.value_counts()
                         average_ratings = df.groupby(['movieId']).rating.mean()
-                        C = average_ratings.mean()  # average rating of a movie
+                        C = average_ratings.mean() #average rating of a movie
                         q_movies = df.movieId.unique()
                         unique_val_counts = df.movieId.value_counts()
-                        q_movies = [row for row in q_movies if unique_val_counts[row] > m]
-
-                        def weighted_rating(movie_id, m=m, C=C):
+                        q_movies = [row for row in q_movies if unique_val_counts[row]>m]
+                        def weighted_rating(movie_id,m=m,C=C):
                             v = vote_count[movie_id]
                             R = average_ratings[movie_id]
-                            weighted_score = (v * R / (v + m)) + (m * C / (v + m))
-                            return (round(weighted_score, 2))
-
+                            weighted_score = (v*R/(v+m))+(m*C/(v+m))
+                            return(round(weighted_score,2))
                         def Top_N_Recommendations(movies_list=q_movies):
                             weighted_scores = [weighted_rating(movie_id) for movie_id in movies_list]
-                            trending_df = pd.DataFrame({'movieId': movies_list, 'IMDB_score': weighted_scores})
-                            trending_df = trending_df.sort_values('IMDB_score', ascending=False)
-                            trending_df = trending_df.merge(data_f1, on=['movieId'], how='inner')
-                            return (trending_df)
-
+                            trending_df = pd.DataFrame({'movieId':movies_list,'IMDB_score':weighted_scores})
+                            trending_df = trending_df.sort_values('IMDB_score',ascending=False)
+                            trending_df = trending_df.merge(data_f1, on=['movieId'],how = 'inner')
+                            return(trending_df)
                         trending_df = Top_N_Recommendations(movies_list=q_movies)
                         trending_df.index = np.arange(1, len(trending_df) + 1)
-                        return (trending_df.head(10)[['title', 'release_year', 'IMDB_score']])
-
-                    return (top_movies_list())
-
+                        return(trending_df.head(10)[['title','release_year','IMDB_score']])
+                    return(top_movies_list())
                 with st.spinner('Let\'s look back through time...'):
                     st.table(average_by_year(0))
 
-        if drop_down_listings == 'Movie Search':
+        if drop_down_listings =='Movie Search':
             value = st.text_input('Movie title', '')
-            actor = st.text_input('Actor or producer:', '')
+            actor = st.text_input('Actor or producer:','')
             indexes = []
             count = 0
             chunksize = 100000
-            chunks = [title_list[x:x + chunksize] for x in range(0, len(title_list), chunksize)]
+            chunks = [title_list[x:x+chunksize] for x in range(0, len(title_list), chunksize)]
             for chunk in chunks:
-                for index, title in enumerate(chunk):
+                for index,title in enumerate(chunk):
                     if value.lower() in title.lower():
-                        indexes.append(count + index)
+                        indexes.append(count+index)
                     else:
                         pass
                 count += chunksize
                 options = np.array(title_list)[indexes]
             actors_and_dir = pd.DataFrame()
-            for chunk_imdb in pd.read_csv('../unsupervised_data/unsupervised_movie_data/imdb_data.csv',
-                                          chunksize=100000):
-                chunk = chunk_imdb[(chunk_imdb.title_cast.map(
-                    lambda x: actor.lower() in str(x).lower()) | chunk_imdb.director.map(
-                    lambda x: actor.lower() in str(x).lower()))]
-                actors_and_dir = pd.concat([chunk, actors_and_dir], axis=1)
+            for chunk_imdb in pd.read_csv('../unsupervised_data/unsupervised_movie_data/imdb_data.csv',chunksize=100000):
+                chunk = chunk_imdb[(chunk_imdb.title_cast.map(lambda x: actor.lower() in str(x).lower())|chunk_imdb.director.map(lambda x: actor.lower() in str(x).lower()))]
+                actors_and_dir = pd.concat([chunk,actors_and_dir],axis=1)
             actors_and_dir = actors_and_dir[['movieId']]
-            for chunk in pd.read_csv('resources/data/movies.csv', chunksize=100000):
-                actors_and_dir = actors_and_dir.merge(chunk, on='movieId', how='left')
-            if len(value) == 0 and len(actor) == 0:
+            for chunk in pd.read_csv('resources/data/movies.csv',chunksize=100000):
+                actors_and_dir = actors_and_dir.merge(chunk,on='movieId',how='left')
+            if len(value)==0 and len(actor)==0:
                 options = ''
-            if len(value) == 0 and len(actor) > 0:
+            if len(value)==0 and len(actor) > 0:
                 options = [title for title in actors_and_dir.title]
-            if len(value) > 0 and len(actor) > 0:
+            if len(value)>0 and len(actor)>0:
                 options = [title for title in actors_and_dir.title if title in options]
             value = st.selectbox("title", options)
             if st.button('view ratings'):
                 def movieId(filename):
-                    chunks = pd.read_csv(filename, chunksize=10000)
+                    chunks = pd.read_csv(filename,chunksize=10000)
                     for chunk in chunks:
                         chunk.title = chunk.title.apply(lambda x: str(x).lower())
-                        if len(chunk[chunk.title == value.lower()]):
-                            return (chunk[chunk.title == value.lower()])
-
+                        if len(chunk[chunk.title==value.lower()]):
+                            return(chunk[chunk.title==value.lower()])
                 selid = movieId('resources/data/movies.csv').movieId.values[0]
-
                 def rate(filename):
-                    chunks = pd.read_csv(filename, chunksize=50000)
+                    chunks = pd.read_csv(filename,chunksize=50000)
                     data = pd.DataFrame()
                     for chunk in chunks:
-                        chunk = chunk[chunk['movieId'] == selid][['movieId', 'rating']]
-                        data = pd.concat([chunk, data])
+                        chunk = chunk[chunk['movieId']==selid][['movieId','rating']]
+                        data = pd.concat([chunk,data])
                     data = data.rating.value_counts()
-                    data = data * 100 / data.sum()
-                    fig = plt.figure(figsize=(25, 10))
+                    data = data*100/data.sum()
+                    fig = plt.figure(figsize=(25,10))
                     ax = fig.add_subplot(111)
-                    labels = [str(round(index, 2)) + ' stars' for index in data.index]
+                    labels = [str(round(index,2))+' stars' for index in data.index]
                     theme = plt.get_cmap('Reds')
                     ax.set_prop_cycle("color", [theme(1. * i / len(labels)) for i in range(len(labels))])
                     sns.set(font_scale=2)
-                    pie = ax.pie(data, autopct='%1.1f%%', shadow=True,
-                                 startangle=10, pctdistance=1.115,
-                                 explode=tuple([0.1 for i in range(len(labels))]))
-                    centre_circle = plt.Circle((0, 0), 0.70, fc='white')
+                    pie = ax.pie(data,autopct='%1.1f%%',shadow=True,
+                                 startangle=10,pctdistance=1.115,
+                                 explode = tuple([0.1 for i in range(len(labels))]))
+                    centre_circle = plt.Circle((0,0),0.70,fc='white')
                     fig = plt.gcf()
                     fig.gca().add_artist(centre_circle)
                     plt.legend(pie[0], labels, loc="lower left")
                     ax.set_title(f'Rating distribution for {value}', fontsize=35)
                     plt.tight_layout()
                     plt.show()
-                    return (ax.figure)
-
+                    return(ax.figure)
                 st.write(rate('../unsupervised_data/unsupervised_movie_data/train.csv'))
 
     if page_selection == "Solution Overview":
+        
         sol_page = """
                     <div class="w3-display-container" style="margin-bottom:15px">
                     <img src="https://www2.cs.duke.edu/courses/fall16/compsci101/assign/assign8-recommender/netflix1.png" style="width:100%">
@@ -416,17 +386,15 @@ def main():
                     <h2><b>Solution Overview</b></h2>
                     </div>
                     </div>
-
                     <div class="w3-row-padding w3-content" style="max-width:100%;">
                     <div style="padding-left: 0;">
                     <div class="w3-justify">
-                    <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue">
-                    <p style="padding-top: 10px;width: 100%;"><b>The Solution</b>></p>
-                    </div>
-                    
-                    <p><img src="https://www.scielo.br/img/revistas/jistm/v13n3//1807-1775-jistm-13-03-0497-gf01.jpg" alt="" height="320" width="280" style="float: right;">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus imperdiet, nulla et dictum interdum, nisi lorem egestas odio, vitae scelerisque enim ligula venenatis dolor. Maecenas nisl est, ultrices nec congue eget, auctor vitae massa. Fusce luctus vestibulum augue ut aliquet. Mauris ante ligula, facilisis sed ornare eu, lobortis in odio. Praesent convallis urna a lacus interdum ut hendrerit risus congue. Nunc sagittis dictum nisi, sed ullamcorper ipsum dignissim ac. In at libero sed nunc venenatis imperdiet sed ornare turpis. Donec vitae dui eget tellus gravida venenatis. Integer fringilla congue eros non fermentum. Sed dapibus pulvinar nibh tempor porta. Cras ac leo purus. Mauris quis diam velit.</p>
                     <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue"  style="margin-top: 70px;">
-                    <p style="padding-top: 10px;"><b>How it Works</b>></p>
+                    <p style="padding-top: 10px;"><b>The Solution</b></p>
+                    </div>
+                    <p><img src="https://www.scielo.br/img/revistas/jistm/v13n3//1807-1775-jistm-13-03-0497-gf01.jpg" alt="" height="320" width="280" style="float: right;">Build a Collaborative filtering recommendation engine. Based on a users' historic movie data we first predict how a user will rate a movie they have not yet rated, once we have the estimated ratings, we then recommend the topN highly rated(estimated ratings) movies. To arrive at the best rating estimates, <b>RMSE</b>(root mean squared error) is the cost function to be minimized, we want to achieve an RMSE score less than or equal <b>0.85</b>. <br>The fundamental assumption behind <b>collaborative filtering</b> technique is that similar user preferences over the items could be exploited to recommend those items to a user who has not seen or used it before. In simpler terms, we assume that users who agreed in the past (viewed the same movie) will agree in the future.<br></p>
+                    <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue"  style="margin-top: 70px;">
+                    <p style="padding-top: 10px;"><b>How it Works</b></p>
                     </div>
                     <h3>Model based methods</h3>
                     <p>Model based collaborative approaches only rely on user-item interactions information and assume a latent model supposed to explain these interactions. For example, matrix factorisation algorithms consists in decomposing the huge and sparse user-item interaction matrix into a product of two smaller and dense matrices: a user-factor matrix (containing users representations) that multiplies a factor-item matrix (containing items representations).</p>
@@ -440,19 +408,13 @@ def main():
                     <li>VT(transpose) is a right singular orthogonal matrix, indicating the similarity between items and latent factors.</p></li>
                     </ul>
                     <img src="https://miro.medium.com/max/697/1*bFmQ2DzGokaBUssGkFR3gg.png" alt="" height="250" width="450">
-                    <div class="w3-panel w3-pale-blue w3-leftbar w3-rightbar w3-border-blue">
-                    <p style="padding-top: 10px;"><b>The Advantages</b>></p>
-                    </div>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
                     </div>
                     </div>
                     </div>
                     </div>
                     """
         st.markdown(sol_page, unsafe_allow_html=True)
-
-    # You may want to add more sections here for aspects such as an EDA,
-    # or to provide your business pitch.
+    
     if page_selection == "About us":
         the_team = """
                     <h1 style="color:black;padding-bottom: 15px"><b>Meet the team that made it happen</b></h1>
@@ -532,6 +494,10 @@ def main():
                     </div>
                     """
         st.markdown(the_team,unsafe_allow_html=True)
+
+
+    # You may want to add more sections here for aspects such as an EDA,
+    # or to provide your business pitch.
 
 
 if __name__ == '__main__':
